+++ src/ui/templates/base.html
@@ -35,7 +35,7 @@
     </div>
     <script
       type="text/javascript"
-      src="{{ url_for('static', filename='dist/js/projects.min.js') }}"
+      src="{{ url_for('static', filename='dist/js/index.min.js') }}"
     ></script>
   </body>
 </html>
